requirements_dev.txt
autopep8==1.6.0
flake8==4.0.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-bitcoinrpc==1.0
python-dotenv==0.20.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
